refactor(content_ui): localize website graphql
