+++ frontend/plugins/content_ui/src/modules/cms/categories/components/CategoriesNavigation.tsx
@@ -10,7 +10,10 @@ import { PageHeader } from 'ui-modules';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useQuery } from '@apollo/client';
-import { CONTENT_CMS_LIST, GET_CLIENT_PORTALS } from '../../graphql/queries';
+import {
+  CONTENT_CMS_LIST,
+  GET_CLIENT_PORTALS,
+} from '@/cms/websites/graphql';
 
 export const CategoriesNavigation = () => {
   const { t } = useTranslation('content');
