+++ frontend/plugins/content_ui/src/modules/cms/websites/graphql/index.ts
@@ -1,0 +1,10 @@
+export {
+  CONTENT_CMS_LIST,
+  GET_CLIENT_PORTALS,
+  GET_WEBSITES,
+} from './queries';
+export {
+  CONTENT_CREATE_CMS,
+  CONTENT_DELETE_CMS,
+  CONTENT_UPDATE_CMS,
+} from './mutations';
